rename test classes (containsOnlyInstancesOf -> hasOnlyElementsOfTypes)
